requirements.txt
# Core dependencies
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.8.0
scipy==1.10.1
joblib==1.4.2

# XAI libraries
lime==0.2.0.1
eli5==0.10.1

# Web framework
fastapi==0.115.12
uvicorn==0.33.0
python-multipart==0.0.20
starlette>=0.40.0
werkzeug==2.2.3
jinja2==3.1.2
markupsafe>=2.1.1

# Testing
pytest==8.3.5
requests==2.32.3
pluggy==1.5.0
tomli==2.2.1
iniconfig==2.1.0

# Python utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
threadpoolctl==3.5.0
importlib-resources==6.4.5
zipp==3.20.2
importlib-metadata==8.5.0
python-dotenv==1.0.1
loguru==0.7.2

# TensorFlow dependencies
absl-py==2.2.2
tensorboard==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
keras==2.8.0
h5py==3.1.0
google-auth==2.28.2
markdown==3.5.2
protobuf==3.20.0
grpcio>=1.24.3
six==1.16.0
